Build 3878: non-blicking dialogs, 'Inhibit TB activity', disable 'Reset primary interface...', and fix compiler warnings:
 * Changes many warning dialogs so they do not block other Tunnelblick operations.

 * Includes a new 'Inhibit automatic update checking and IP Address checking' checkbox on the 'Preferences' page. Checking it disables all Internet activity by the Tunnelblick program itself (but not by OpenVPN), overriding any other settings that allow such activity. The checkbox is unchecked by default.

 * Disables the 'Reset the primary interface after disconnecting' checkbox if 'Set DNS/WINS' is not set to 'Set nameserver'.

 * Includes changes to avoid compiler warnings on Xcode 5 and 6.

+++ trunk/tunnelblick/openvpnstart.m
@@ -1939,7 +1939,7 @@ void printSanitizedConfigurationFile(NSString * configFile, unsigned cfgLocCode)
             if (  [tokens count] > 0  ) {
                 NSString * firstToken = [tokens objectAtIndex: 0];
                 if (  [firstToken hasPrefix: @"<"]  ) {
-                    unsigned j;
+                    NSUInteger j;
                     if (  (j = [beginInlineKeys indexOfObject: firstToken]) != NSNotFound  ) {
                         unsigned beginLineNumber = i;
 						BOOL foundEnd = FALSE;
